moved model and instance methods to model
signture from env
