+++ server/controllers/user.controllers.js
@@ -33,6 +33,7 @@ export const loginUser = async (req, res) => {
     const token = await user.generateAuthToken();
     res.status(200).send({ user, token });
   } catch (error) {
+    console.log(error);
     res.status(401).send(error);
   }
 };
@@ -71,5 +72,3 @@ export const editProfile = async (req, res) => {
     res.status(401);
   }
 };
-
-nvjbhjkbhj;
