+++ server/middleware/auth/auth.js
@@ -1,10 +1,11 @@
 import jwt from "jsonwebtoken";
+import "dotenv/config";
 import { User } from "../../models/user/user.model.js";
 
 export const auth = async (req, res, next) => {
   try {
     const token = req.header("Authorization").replace("Bearer ", "");
-    const decoded = jwt.verify(token, "hackathonGroup2");
+    const decoded = jwt.verify(token, process.env.JWT_TOKEN_SECRET);
     const user = await User.findOne({
       _id: decoded._id,
       "tokens.token": token,
