+++ server/models/user/user.schema.js
@@ -1,8 +1,5 @@
 import mongoose from "mongoose";
 import validator from "validator";
-import bcrypt from "bcrypt";
-import jwt from "jsonwebtoken";
-import { User } from "./user.model.js";
 
 const userSchema = new mongoose.Schema({
   name: {
@@ -34,42 +31,6 @@ const userSchema = new mongoose.Schema({
       },
     },
   ],
-});
-
-userSchema.methods.toJSON = function () {
-  const user = this;
-  const userObject = user.toObject();
-  delete userObject.tokens;
-  delete userObject.password;
-
-  return userObject;
-};
-
-userSchema.methods.generateAuthToken = async function (params) {
-  const user = this;
-  const token = jwt.sign({ _id: user._id.toString() }, "hackathonGroup2");
-
-  user.tokens = user.tokens.concat({ token });
-
-  await user.save();
-  return token;
-};
-
-userSchema.statics.findByCredentials = async (email, password) => {
-  const user = await User.findOne({ email });
-  if (!user) throw new Error("Something went wrong");
-  const isMatch = await bcrypt.compare(password, user.password);
-  if (!isMatch) throw new Error("Something went wrong");
-
-  return user;
-};
-
-userSchema.pre("save", async function (next) {
-  const user = this;
-  if (user.isModified("password")) {
-    user.password = await bcrypt.hash(user.password, 12);
-  }
-  next();
 });
 
 export { userSchema };
